Move verify_opod into sentinel1.orbits
