+++ hyp3lib/depreciated/verify_opod.py
@@ -4,31 +4,7 @@ import argparse
 import logging
 import os
 
-from lxml import etree
-
-
-def verify_opod(fi):
-    logging.info("Verifying state vector file")
-    root = etree.parse(fi)
-    check = 0
-    for item in root.iter('File_Description'):
-        if "Orbit File" not in item.text:
-            raise ValueError("Not an orbit file!")
-        else:
-            logging.info("...Found orbit file")
-            check += 1
-    for item in root.iter('File_Type'):
-        if "AUX_POEORB" not in item.text and "AUX_PREORB" not in item.text and "AUX_RESORB" not in item.text:
-            raise ValueError("Unknown file type!")
-        else:
-            logging.info("...Found file type {}".format(item.text))
-            check += 1
-
-    if not check:
-        raise ValueError("Not a valid state vector file: {}".format(fi))
-
-    else:
-        logging.info("State vector file verified")
+from hyp3lib.sentinel1.orbits import verify_opod
 
 
 def main():
